relocate the `common` directory
