+++ src/main/java/io/github/wmdietl/diagnostics/common/Diagnostic.java
@@ -1,4 +1,4 @@
-package io.github.wmdietl.diagnostics.json.common;
+package io.github.wmdietl.diagnostics.common;
 
 // Interface for multiple types of Diagnostics (currently: javac, lsp, sarif)
 public interface Diagnostic {}
