+++ src/main/java/io/github/wmdietl/diagnostics/json/javac/JavacDiagnostic.java
@@ -1,7 +1,8 @@
 package io.github.wmdietl.diagnostics.json.javac;
 
 import javax.tools.JavaFileObject;
-import io.github.wmdietl.diagnostics.json.common.Diagnostic;
+
+import io.github.wmdietl.diagnostics.common.Diagnostic;
 
 public class JavacDiagnostic implements Diagnostic {
