+++ src/main/java/io/github/wmdietl/diagnostics/json/javac/JavacDiagnosticList.java
@@ -4,8 +4,9 @@ import java.util.ArrayList;
 import java.util.List;
 
 import javax.tools.JavaFileObject;
-import io.github.wmdietl.diagnostics.json.common.DiagnosticList;
-import io.github.wmdietl.diagnostics.json.common.Diagnostic;
+
+import io.github.wmdietl.diagnostics.common.Diagnostic;
+import io.github.wmdietl.diagnostics.common.DiagnosticList;
 
 public class JavacDiagnosticList extends DiagnosticList {
     public JavacDiagnosticList(List<javax.tools.Diagnostic<? extends JavaFileObject>> diags) {
