+++ src/main/java/io/github/wmdietl/diagnostics/json/lsp/LspDiagnostic.java
@@ -4,9 +4,9 @@ import java.util.List;
 
 import javax.tools.JavaFileObject;
 
-import io.github.wmdietl.diagnostics.json.common.Range;
-import io.github.wmdietl.diagnostics.json.common.DiagnosticSeverity;
-import io.github.wmdietl.diagnostics.json.common.Diagnostic;
+import io.github.wmdietl.diagnostics.common.Diagnostic;
+import io.github.wmdietl.diagnostics.common.DiagnosticSeverity;
+import io.github.wmdietl.diagnostics.common.Range;
 
 /** 
  * A file plus a list of diagnostics for that file.
