+++ test.java
@@ -1,0 +1,6 @@
+/**
+ * test
+ */
+public class test {
+    Something is Wrong with this file!
+}
